// src/app/shared/model/common-setting/office-bank-acc.model.ts
// Generated by https://quicktype.io

export class OfficeBankAccObj {
    OfficeBankAccId: any;
    RefOfficeId: any;
    OfficeBankAccCode: any;
    OfficeBankAccBranch: any;
    OfficeBankAccName: any;
    OfficeBankAccNo: any;
    IsLegalDoc: any;
    RefBankId: any;
    RefCurrId: any;
    MrBankAccPurposeCode: any;
    BankBranchRegRptCode: any;
    BankAddr: any;
    BankCity: any;
    BankAreaCode1: any;
    BankAreaCode2:any;
    BankAreaCode3: any;
    BankAreaCode4: any;
    BankZipcode:any;
    BankPhnArea1: any;
    BankPhn1: any;
    BankPhnExt1: any;
    BankPhnArea2: any;
    BankPhn2: any;
    BankPhnExt2: any;
    BankPhnArea3: any;
    BankPhn3: any;
    BankPhnExt3: any;
    BankFaxArea: any;
    BankFax: any;
    CntctPersonJobTitle: any;
    CntctPersonName: any;
    BankAccType: any;
    IsActive: any;
    RowVersion: any;
    constructor() { this.OfficeBankAccId = 0, this.RowVersion = "" }
}